requirements.txt
asgiref==3.11.0
Django==4.2
django-filter==25.2
django-jazzmin==2.6.0
django-redis==6.0.0
djangorestframework==3.16.1
pillow==12.0.0
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.4
